# Speed up deploys with -auto-approve