--- a/stream_alert_cli/manage_lambda/deploy.py
+++ b/stream_alert_cli/manage_lambda/deploy.py
@@ -150,4 +150,4 @@ def deploy(options, config):
         return
 
     # Apply the changes to the Lambda aliases
-    helpers.tf_runner(targets=deploy_targets)
+    helpers.tf_runner(targets=deploy_targets, refresh=False, auto_approve=True)