@jacknagz 2018-02-21
@austinbyers 

```
W:132,24: Use % formatting in logging functions and pass the % parameters as arguments (logging-format-interpolation)
```